Revert "refactor: unify robot telemetry and DRY dashboard/control layout"
